+++ src/aideas/app/app.py
@@ -1,16 +1,14 @@
 import pickle
 import shutil
 import logging
-import sys
-from enum import Enum, unique
-from typing import Union, Callable, TypeVar
+from typing import Union
 
 from .agent.agent_factory import AgentFactory
-from .env import Env, get_cached_results_file, get_value
-from pyu.io.file import create_file
+from .env import get_cached_results_file
 from .result.result_set import AgentResultSet, StageResultSet
 from .config_loader import ConfigLoader
 from .run_context import RunContext
+from pyu.io.file import create_file
 
 logger = logging.getLogger(__name__)
 
@@ -39,9 +37,9 @@ class App:
     def title(self):
         return self.__config["app"]["title"]
 
-    def run(self, agent_names: Union[str, list[str], None] = None) -> AgentResultSet:
+    def run(self, run_config: dict[str, any]) -> AgentResultSet:
 
-        run_context: RunContext = RunContext.of_config(self.__config, agent_names)
+        run_context: RunContext = RunContext.of_config(self.__config, run_config)
 
         logger.debug(f"Running agents: {run_context.get_agent_names()}")
 
@@ -76,56 +74,3 @@ class App:
         config_loader: ConfigLoader = self.__agent_factory.get_config_loader()
         config_path = config_loader.get_agent_config_path(agent_name)
         shutil.copy2(config_path, f'{object_path[:object_path.index(name)]}config.yaml')
-
-
-"""
-The following code is used to parse command line arguments.
-"""
-T = TypeVar("T", bound=any)
-
-
-@unique
-class AppArg(str, Enum):
-    def __new__(cls, value, env_name: Env = None):
-        obj = str.__new__(cls, [value])
-        obj._value_ = value
-        obj.__env_name = env_name
-        return obj
-
-    @property
-    def alias(self) -> str:
-        return self._value_
-
-    @property
-    def env_name(self) -> Union[Env, None]:
-        return self.__env_name
-
-    AGENTS = ('a', Env.AGENTS)
-
-
-def get_list_arg_value(arg_name: AppArg) -> [str]:
-    return __get_formatted_arg(arg_name, lambda x: x.split(','), [])
-
-
-def __get_formatted_arg(arg: AppArg,
-                        convert: Callable[[str], T],
-                        result_if_none: Union[T, None] = None) -> T:
-    arg_value = get_arg_value(arg, None)
-    return result_if_none if arg_value is None else convert(arg_value)
-
-
-def get_arg_value(arg: AppArg, result_if_none: Union[any, None] = None) -> any:
-    """
-    Get the value of the argument with the given name.
-    Arguments have aliases that can be used to refer to them.
-    --agents twitter could be written as -a twitter
-    :param arg: The name of the argument.
-    :param result_if_none: The result to return if none
-    :return: The value of the argument with the given name.
-    """
-    args: [str] = [e.lower() for e in sys.argv]
-    candidates: [str] = [f'--{arg.name.lower()}', f'-{arg.alias.lower()}']
-    for candidate in candidates:
-        if candidate in args:
-            return args[args.index(candidate) + 1]
-    return result_if_none if not arg.env_name else get_value(arg.env_name, result_if_none)