+++ src/test/app/agent/blog_agent_test.py
@@ -2,7 +2,7 @@ import logging.config
 import unittest
 
 from test.app.agent.test_blog_agent import TestBlogAgent
-from test.app.test_functions import init_logging, get_config_loader
+from test.app.test_functions import init_logging, load_agent_config, get_run_context
 from aideas.app.agent.agent_name import AgentName
 from aideas.app.run_context import RunContext
 
@@ -11,11 +11,10 @@ init_logging(logging.config)
 
 class BlogAgentTest(unittest.TestCase):
     def test_run(self):
-        app_config = get_config_loader().load_app_config()
         agent_name = AgentName.BLOG
-        run_context: RunContext = RunContext.of_config(app_config, agent_name)
-        agent_config = get_config_loader().load_agent_config(agent_name)
-        agent = TestBlogAgent.of_config(agent_name, app_config, agent_config)
+        run_context: RunContext = get_run_context([agent_name])
+        agent = TestBlogAgent.of_config(
+            agent_name, run_context.get_app_config(), load_agent_config(agent_name))
 
         result = agent.run(run_context)