+++ src/test/app/app_it.py
@@ -1,7 +1,7 @@
 import logging.config
 import unittest
 
-from test_functions import get_config_loader, init_logging
+from test_functions import get_main_config_loader, init_logging, load_run_config
 from agent.test_agent_factory import TestAgentFactory
 from aideas.app.app import App
 
@@ -10,7 +10,7 @@ init_logging(logging.config)
 
 class AppIT(unittest.TestCase):
     def test_run(self):
-        result = given_app().run()
+        result = given_app().run(load_run_config())
         print(f'{result.pretty_str()}')
         self.assertTrue(result.is_successful())
 
@@ -18,8 +18,9 @@ class AppIT(unittest.TestCase):
 def given_app() -> App:
     # This will result to the live app, which we don't want
     #return App.of_defaults(get_main_path('config'))
-    agent_factory = TestAgentFactory(get_config_loader())
-    return App(agent_factory, get_config_loader().load_app_config())
+    config_loader = get_main_config_loader()
+    agent_factory = TestAgentFactory(config_loader)
+    return App(agent_factory, config_loader.load_app_config())
 
 
 if __name__ == '__main__':